+++ src/index.html
@@ -32,7 +32,7 @@
 
     <!-- Contenedor de botones -->
     <div class="container_botones_principal" id="container_botones_principal">
-      <!-- Historia de usuario 2 (filtro)-->
+      <!-- --------------------------- HISTORIA DE USUARIO 2 (filtro)--------------------------- -->
       <select
         id="lista-desplegable-filtrar"
         style="display: block"
@@ -60,11 +60,11 @@
         <option value="fairy">Fairy</option>
       </select>
 
-      <!-- Historia de usuario 3 (orden)-->
+      <!-- --------------------------- HISTORIA DE USUARIO 3 (orden)--------------------------- -->
       <button>Order</button>
     </div>
 
-    <!-- Historia de usuario 1 (contenedor de pokemones) -->
+    <!-- --------------------------- HISTORIA DE USUARIO 1 (visualización)--------------------------- -->
     <div id="contenedor_pokemones" class="contenedor_pokemones"></div>
 
     <script src="main.js" type="module"></script>
